requirements.txt
# GridAlpha - PJM Electricity Market Intelligence Dashboard
# Python 3.10+

# Web framework
streamlit>=1.32.0

# HTTP client
requests>=2.31.0
httpx>=0.27.0

# Data processing
pandas>=2.2.0
numpy>=1.26.0

# Environment variables
python-dotenv>=1.0.0

# Date/time handling
python-dateutil>=2.9.0

# Data caching and serialization
pyarrow>=15.0.0

# Plotting (for future frontend phases)
plotly>=5.20.0
altair>=5.3.0

# Logging
loguru>=0.7.2

# Type hints support
typing-extensions>=4.10.0
